data-pipeline/requirements.txt
# Environment management
python-dotenv>=1.0.0

# NASA EarthData access
earthaccess>=0.0.8

# Data handling
pandas>=2.0
numpy>=1.26
xarray>=2023.12
netCDF4>=1.6

# Optional: for HDF5 files
h5py>=3.9

# For ML preprocessing / merging
scikit-learn>=1.3

# Enhanced data access
requests-oauthlib>=1.3.0
opencage>=1.2.0  # Geocoding
geopy>=2.3.0     # Geographic calculations

# Additional data formats
openpyxl>=3.1.0  # Excel files
beautifulsoup4>=4.12.0  # Web scraping

# Enhanced ML capabilities
lightgbm>=4.0.0  # Alternative to XGBoost
optuna>=3.3.0    # Hyperparameter optimization

# Visualization and analysis
plotly>=5.17.0
seaborn>=0.12.0
folium>=0.14.0   # Interactive maps

# Progress tracking
tqdm>=4.65.0

# Additional utilities
python-dateutil>=2.8.0
pytz>=2023.3